admin update, activate, deactivate Service-category
