+++ Washouse.Web/Controllers/AdminController.cs
@@ -37,11 +37,12 @@ namespace Washouse.Web.Controllers
         private readonly ICenterRequestService _centerRequestService;
         private readonly IFeedbackService _feedbackService;
         private readonly IPostService _postService;
+        private readonly IServiceCategoryService _serviceCategoryService;
         public AdminController(ICenterService centerService, ICloudStorageService cloudStorageService, IAccountService accountService,
                                 ILocationService locationService, IWardService wardService,
                                 IOperatingHourService operatingHourService, IServiceService serviceService,
                                 IStaffService staffService, ICenterRequestService centerRequestService, IFeedbackService feedbackService,
-                                IPostService postService)
+                                IPostService postService, IServiceCategoryService serviceCategoryService)
         {
             this._centerService = centerService;
             this._accountService = accountService;
@@ -52,6 +53,7 @@ namespace Washouse.Web.Controllers
             this._serviceService = serviceService;
             this._staffService = staffService;
             this._centerRequestService = centerRequestService;
+            this._serviceCategoryService = serviceCategoryService;
             this._feedbackService = feedbackService;
             this._postService = postService;
         }
@@ -406,6 +408,112 @@ namespace Washouse.Web.Controllers
 
 
             }
+            
+            
+        }
+           [HttpPut("service-categories/{id}")]
+        public async Task<IActionResult> Update([FromForm] CategoryRequestModel category, int id)
+        {
+            try
+            {
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest(new ResponseModel
+                    {
+                        StatusCode = StatusCodes.Status400BadRequest,
+                        Message = "Model is not valid",
+                        Data = null
+                    });
+                }
+                else
+                {
+                    Category existingCategorySevice = await _serviceCategoryService.GetById(id);
+                    //Category existingCategorySevice = new Category();
+                    if (existingCategorySevice == null)
+                    {
+                        return NotFound(new ResponseModel
+                        {
+                            StatusCode = StatusCodes.Status404NotFound,
+                            Message = "Not found category.",
+                            Data = null
+                        });
+                    }
+                    else
+                    {
+                        existingCategorySevice.CategoryName = category.CategoryName;
+                        existingCategorySevice.Description = category.Description;
+                        existingCategorySevice.UpdatedDate = DateTime.Now;
+                        existingCategorySevice.UpdatedBy = User.FindFirst(ClaimTypes.Email)?.Value;
+                        existingCategorySevice.Status = category.Status;
+                        existingCategorySevice.HomeFlag = category.HomeFlag;
+                        existingCategorySevice.Image = category.SavedFileName;
+                        await _serviceCategoryService.Update(existingCategorySevice);
+                        var response = new CategoryResponseModel
+                        {
+                            CategoryId = existingCategorySevice.Id,
+                            CategoryName = existingCategorySevice.CategoryName,
+                            CategoryAlias = existingCategorySevice.Alias,
+                            Description = existingCategorySevice.Description,
+                            HomeFlag = existingCategorySevice.HomeFlag,
+                            Image = existingCategorySevice.Image != null
+                                ? await _cloudStorageService.GetSignedUrlAsync(existingCategorySevice.Image)
+                                : null,
+                        };
+                        return Ok(new ResponseModel
+                        {
+                            StatusCode = 0,
+                            Message = "success",
+                            Data = response
+                        });
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new ResponseModel
+                {
+                    StatusCode = StatusCodes.Status400BadRequest,
+                    Message = ex.Message,
+                    Data = null
+                });
+            }
+        }
+
+        [HttpPut("service-categories/{id}/deactivate")]
+        public async Task<IActionResult> DeactivateCategory(int id)
+        {
+            var category = await _serviceCategoryService.GetById(id);
+            if (category == null)
+            {
+                return NotFound();
+            }
+
+            var services = _serviceService.GetServicesByCategory(category.Id);
+            if (services.Any(serivce => serivce.Status.ToLower().Equals("active")))
+            {
+                return BadRequest(new ResponseModel
+                {
+                    StatusCode = StatusCodes.Status400BadRequest,
+                    Message = "Deactivate failed due to some active services",
+                    Data = null
+                });
+            }
+
+            await _serviceCategoryService.DeactivateCategory(id);
+            return Ok();
+        }
+
+        [HttpPut("service-categories/{id}/activate")]
+        public async Task<IActionResult> ActivateCategory(int id)
+        {
+            var category = await _serviceCategoryService.GetById(id);
+            if (category == null)
+            {
+                return NotFound();
+            }
+
+            await _serviceCategoryService.ActivateCategory(id);
+            return Ok();
         }
     }
 }
